+++ docs/kr/ac/snu/sbkim28/game/core/ConsoleInputPlayer.html
@@ -73,7 +73,7 @@ loadScripts(document, 'script');</script>
 </div>
 <div class="inheritance" title="Inheritance Tree"><a href="https://docs.oracle.com/en/java/javase/17/docs/api/java.base/java/lang/Object.html" title="class or interface in java.lang" class="external-link">java.lang.Object</a>
 <div class="inheritance"><a href="AbstractPlayer.html" title="class in kr.ac.snu.sbkim28.game.core">kr.ac.snu.sbkim28.game.core.AbstractPlayer</a>
-<div class="inheritance">kr.ac.snu.sbkim28.game.io.ConsoleInputPlayer</div>
+<div class="inheritance">kr.ac.snu.sbkim28.game.dp.ConsoleInputPlayer</div>
 </div>
 </div>
 <section class="class-description" id="class-description">
